requirements.txt
requests
bs4
PyPDF2
python-dotenv
datasets
openai
anthropic
